添加Action的基础

+++ src/Practice/CozyKxlol/Kxlol/Object/CozyCircle.cs
@@ -256,9 +256,9 @@ namespace CozyKxlol.Kxlol.Object
 
             if (HasBorder)
             {
-                spriteBatch.DrawCircle(center, Radius + BorderSize, (int)Radius, Color.Black, BorderSize);
+                spriteBatch.DrawCircle(center, Radius + BorderSize, Radius, Color.Black, BorderSize);
             }
-            spriteBatch.DrawCircle(center, Radius, (int)Radius, ColorProperty, Radius);
+            spriteBatch.DrawCircle(center, Radius, Radius, ColorProperty, Radius);
         }
 
         // IMoveAble
